@@ -17,18 +17,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.2.3",
-    "@angular/cdk": "10.2.7",
-    "@angular/common": "10.2.3",
-    "@angular/compiler": "10.2.3",
-    "@angular/core": "10.2.3",
-    "@angular/forms": "10.2.3",
-    "@angular/localize": "10.2.3",
-    "@angular/material": "10.2.7",
-    "@angular/platform-browser": "10.2.3",
-    "@angular/platform-browser-dynamic": "10.2.3",
-    "@angular/router": "10.2.3",
-    "@auth0/angular-jwt": "5.0.1",
+    "@angular/animations": "11.0.2",
+    "@angular/cdk": "11.0.0",
+    "@angular/common": "11.0.2",
+    "@angular/compiler": "11.0.2",
+    "@angular/core": "11.0.2",
+    "@angular/forms": "11.0.2",
+    "@angular/localize": "11.0.2",
+    "@angular/material": "11.0.0",
+    "@angular/platform-browser": "11.0.2",
+    "@angular/platform-browser-dynamic": "11.0.2",
+    "@angular/router": "11.0.2",
+    "@auth0/angular-jwt": "5.0.2",
     "@fortawesome/angular-fontawesome": "0.8.0",
     "@fortawesome/fontawesome-svg-core": "1.2.32",
     "@fortawesome/free-brands-svg-icons": "5.15.1",
@@ -58,23 +58,24 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "0.1100.2",
-    "@angular/cli": "10.2.0",
-    "@angular/compiler-cli": "10.2.3",
-    "@angular/language-service": "10.2.3",
+    "@angular/cli": "11.0.2",
+    "@angular/compiler-cli": "11.0.2",
+    "@angular/language-service": "11.0.2",
     "@ngrx/schematics": "10.0.1",
-    "@types/jasmine": "3.6.1",
+    "@types/jasmine": "3.6.2",
     "@types/jasminewd2": "2.0.8",
+    "@types/lodash": "4.14.165",
     "@types/node": "13.13.30",
     "@types/ua-parser-js": "0.7.33",
     "codelyzer": "6.0.1",
     "jasmine-core": "3.6.0",
-    "jasmine-spec-reporter": "6.0.0",
+    "jasmine-spec-reporter": "5.0.2",
     "karma": "5.2.3",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-coverage": "2.0.3",
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.5.4",
-    "prettier": "2.1.2",
+    "prettier": "2.2.0",
     "protractor": "7.0.0",
     "ts-node": "9.0.0",
     "tslint": "6.1.3",
