@@ -1,135 +1,64 @@
-import { Component, OnInit, OnDestroy, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
+import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { takeUntil, map } from 'rxjs/operators';
-import { Subject, Observable, combineLatest } from 'rxjs';
+import { map, tap, filter } from 'rxjs/operators';
+import { Observable, combineLatest } from 'rxjs';
 
 import { Store, select } from '@ngrx/store';
-import { cloneDeep } from 'lodash';
 
 import { AppState } from '@/core/reducers';
 import { currentUser, isPrivileged } from '@/core/auth/store/auth.selectors';
-import { EDIT_PRIVILEGES, ACTION_LABELS, CONSTANTS, MAT_BUTTON, RoutingDataConstants } from '@/shared/constants';
-import { DialogType, DynamicForm } from '@/shared/models';
-import { UsersDialogComponent, User } from '@/modules/users';
+import { EDIT_PRIVILEGES, FORMCONSTANTS } from '@/shared/constants';
 import { Department } from '../../models';
-import { updateDepartmentRequested, changeIsEditingState } from '../../store/department.actions';
-import { selectIsEditing } from '../../store/department.selectors';
-import { DialogConfirmModel } from '@/shared/models/dialog/dialog-confirm.model';
-import { DialogDataModel, DialogWithTwoButtonModel, DialogResultModel } from '@/shared/models';
-import { DialogConfirmComponent } from '@/shared/components/dialogs/dialog-confirm/dialog-confirm.component';
-import { DialogService } from '@/shared/services';
+import { updateDepartmentRequested, changeIsEditingState, departmentRequested } from '../../store/department.actions';
+import { selectDepartmentById, selectIsEditing } from '../../store/department.selectors';
 
 @Component({
-  templateUrl: './department.component.html',
+  template: `
+    <templates-department-view
+      [editForm]="editForm$ | async"
+      [formName]="formName"
+      [item]="department$ | async"
+      [canEdit]="canEditDepartment$ | async"
+      (saveChanges)="updateDepartment($event)"
+      (isEditing)="onFormStateChange($event)"
+    ></templates-department-view>
+  `,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class DepartmentComponent implements OnInit, OnDestroy {
+export class DepartmentComponent implements OnInit {
   editForm$: Observable<boolean> = this.store$.pipe(select(selectIsEditing));
-  canEditDepartment$: Observable<boolean> = combineLatest([
-    this.store$.pipe(select(isPrivileged(EDIT_PRIVILEGES.DEPARTMENT))),
-    this.store$.pipe(select(currentUser))
-  ]).pipe(map(([editPriv, appUser]) => editPriv || appUser.id === this.department.managerId));
+  department$: Observable<Department>;
+  canEditDepartment$: Observable<boolean>;
 
-  departmentFormJson: DynamicForm;
-  department: Department;
-  departmentInitial: Department;
-  departmentFormValues: Department;
+  formName = FORMCONSTANTS.DEPARTMENT;
 
-  actionLabels = ACTION_LABELS;
-  matButtonTypes = MAT_BUTTON;
-
-  private unsubscribe: Subject<void> = new Subject();
-
-  constructor(
-    private route: ActivatedRoute,
-    private store$: Store<AppState>,
-    private cdr: ChangeDetectorRef,
-    private dialogService: DialogService
-  ) {}
+  constructor(private route: ActivatedRoute, private store$: Store<AppState>) {}
 
   ngOnInit(): void {
-    this.departmentFormJson = this.route.snapshot.data[RoutingDataConstants.FORM_JSON];
-    this.departmentInitial = cloneDeep(this.route.snapshot.data[RoutingDataConstants.DEPARTMENT]);
-    // TODO: @IMalaniak add plan directly from the store
-    this.department = cloneDeep(this.route.snapshot.data[RoutingDataConstants.DEPARTMENT]);
-  }
+    const id: number = this.route.snapshot.params['id'];
 
-  onClickEdit(): void {
-    this.store$.dispatch(changeIsEditingState({ isEditing: true }));
-  }
-
-  onClickCancelEdit(): void {
-    this.store$.dispatch(changeIsEditingState({ isEditing: false }));
-    this.department = cloneDeep(this.departmentInitial);
-  }
-
-  departmentFormValueChanges(formVal: Department): void {
-    this.departmentFormValues = { ...this.departmentFormValues, ...formVal };
-  }
-
-  addManagerDialog(): void {
-    // TODO: @ArseniiIrod, @IMalaniak implement logic with selected user
-    const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
-      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_MANAGER)
-    };
-
-    this.dialogService
-      .open(UsersDialogComponent, dialogDataModel, DialogType.MAX)
-      .afterClosed()
-      .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: DialogResultModel<User[]>) => {
-        if (result && result.success) {
-          this.department = { ...this.department, Manager: { ...result[0] } };
-          this.cdr.detectChanges();
+    this.department$ = this.store$.pipe(
+      select(selectDepartmentById(id)),
+      tap((department) => {
+        if (!department) {
+          this.store$.dispatch(departmentRequested({ id }));
         }
-      });
+      }),
+      filter((department) => !!department)
+    );
+
+    this.canEditDepartment$ = combineLatest([
+      this.store$.pipe(select(isPrivileged(EDIT_PRIVILEGES.DEPARTMENT))),
+      this.store$.pipe(select(currentUser)),
+      this.department$
+    ]).pipe(map(([editPriv, appUser, department]) => editPriv || appUser.id === department.managerId));
   }
 
-  addWorkersDialog(): void {
-    // TODO: @ArseniiIrod, @IMalaniak implement logic with selected users
-    const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
-      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_WORKERS)
-    };
-
-    this.dialogService
-      .open(UsersDialogComponent, dialogDataModel, DialogType.MAX)
-      .afterClosed()
-      .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: DialogResultModel<User[]>) => {
-        if (result && result.success) {
-          const selectedWorkers: User[] = result.model.filter(
-            (selectedWorker) => !this.department.Workers.some((user) => user.id === selectedWorker.id)
-          );
-
-          if (selectedWorkers.length) {
-            this.department.Workers = [...this.department.Workers, ...selectedWorkers];
-            this.cdr.detectChanges();
-          }
-        }
-      });
+  onFormStateChange(isEditing: boolean): void {
+    this.store$.dispatch(changeIsEditingState({ isEditing }));
   }
 
-  removeManager(): void {
-    this.department = { ...this.department, Manager: null };
-  }
-
-  removeWorker(userId: number): void {
-    this.department = { ...this.department, Workers: this.department.Workers.filter((worker) => worker.id !== userId) };
-  }
-
-  updateDepartment(): void {
-    const dialogModel: DialogConfirmModel = new DialogConfirmModel(CONSTANTS.TEXTS_UPDATE_DEPARTMENT_CONFIRM);
-    const dialogDataModel: DialogDataModel<DialogConfirmModel> = { dialogModel };
-
-    this.dialogService.confirm(DialogConfirmComponent, dialogDataModel, () => {
-      this.store$.dispatch(
-        updateDepartmentRequested({ department: { ...this.department, ...this.departmentFormValues } })
-      );
-    });
-  }
-
-  ngOnDestroy(): void {
-    this.unsubscribe.next();
-    this.unsubscribe.complete();
+  updateDepartment(department: Department): void {
+    this.store$.dispatch(updateDepartmentRequested({ department }));
   }
 }
