@@ -3,20 +3,24 @@ import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { RouterStateSerializer, StoreRouterConnectingModule, DefaultRouterStateSerializer } from '@ngrx/router-store';
-import { PreferencesEffects } from './reducers/preferences.effects';
-import { reducers, metaReducers } from './reducers';
-import { httpInterceptorsProviders } from './_interceptors';
-import { LocalStorageService } from '@/shared/services';
+import { StoreRouterConnectingModule, RouterState } from '@ngrx/router-store';
+
 import { CustomSerializer } from '@/shared/utils';
+import { LocalStorageService } from '@/shared/services';
+import { environment } from 'environments/environment';
+import { httpInterceptorsProviders } from './_interceptors';
 import { AuthModule } from './auth/auth.module';
 import { LayoutModule } from './layout/layout.module';
-import { environment } from 'environments/environment';
+import { reducers, metaReducers } from './reducers';
+import { PreferencesEffects } from './reducers/preferences.effects';
 import { IntegrationsEffects } from './reducers/integration.effects';
 import { DynamicFormEffects } from './reducers/dynamic-form/dynamic-form.effects';
+import { DynamicFormService } from './services/dynamic-form.service';
+import { MediaQueryService } from './services';
 
 @NgModule({
   imports: [
@@ -36,13 +40,13 @@ import { DynamicFormEffects } from './reducers/dynamic-form/dynamic-form.effects
     }),
     !environment.production ? StoreDevtoolsModule.instrument() : [],
     EffectsModule.forRoot([PreferencesEffects, IntegrationsEffects, DynamicFormEffects]),
-    StoreRouterConnectingModule.forRoot({ serializer: DefaultRouterStateSerializer, stateKey: 'router' })
+    StoreRouterConnectingModule.forRoot({
+      serializer: CustomSerializer,
+      stateKey: 'router',
+      routerState: RouterState.Full
+    })
   ],
   exports: [RouterModule, HttpClientModule, FormsModule, ReactiveFormsModule, LayoutModule],
-  providers: [
-    httpInterceptorsProviders,
-    LocalStorageService,
-    { provide: RouterStateSerializer, useClass: CustomSerializer }
-  ]
+  providers: [httpInterceptorsProviders, MediaQueryService, DynamicFormService, LocalStorageService]
 })
 export class CoreModule {}
