@@ -2,12 +2,14 @@ import { ActionReducerMap, MetaReducer, createFeatureSelector } from '@ngrx/stor
 import { environment } from 'environments/environment';
 
 import * as fromRouter from '@ngrx/router-store';
+
+import { RouterStateUrl } from '@/shared/utils';
 import * as fromPreferences from './preferences.reducer';
 import * as fromIntegrations from './integration.reducer';
 import * as fromDynamicForm from './/dynamic-form/dynamic-form.reducer';
 
 export interface AppState {
-  router: fromRouter.RouterReducerState<any>;
+  router: fromRouter.RouterReducerState<RouterStateUrl>;
   preferences: fromPreferences.PreferencesState;
   integrations: fromIntegrations.IntegrationsState;
   forms: fromDynamicForm.DynamicFormState;
@@ -22,7 +24,7 @@ export const reducers: ActionReducerMap<AppState> = {
 
 export const metaReducers: MetaReducer<AppState>[] = !environment.production ? [] : [];
 
-export const selectRouter = createFeatureSelector<AppState, fromRouter.RouterReducerState<any>>('router');
+export const selectRouter = createFeatureSelector<AppState, fromRouter.RouterReducerState<RouterStateUrl>>('router');
 
 export const {
   selectCurrentRoute, // select the current route
