@@ -1,22 +1,30 @@
 import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
-import { DepartmentComponent } from './department.component';
+import { HttpClientModule } from '@angular/common/http';
+import { RouterTestingModule } from '@angular/router/testing';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { provideMockStore } from '@ngrx/store/testing';
 import { SharedModule } from '@/shared/shared.module';
-import { HttpClientModule } from '@angular/common/http';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { RouterTestingModule } from '@angular/router/testing';
+import { authStateMock, formsStateMock } from '@/shared/testing/mocks';
+
+import { initialDepartmentsState } from '../../store/department.reducer';
+import { DepartmentComponent } from './department.component';
 
 describe('DepartmentComponent', () => {
   let component: DepartmentComponent;
   let fixture: ComponentFixture<DepartmentComponent>;
+  const initialState = {
+    departments: initialDepartmentsState,
+    auth: authStateMock,
+    forms: formsStateMock
+  };
 
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
         declarations: [DepartmentComponent],
         imports: [RouterTestingModule, BrowserAnimationsModule, HttpClientModule, SharedModule],
-        providers: [provideMockStore({})]
+        providers: [provideMockStore({ initialState })]
       }).compileComponents();
     })
   );
