@@ -6,16 +6,16 @@ export interface Plan {
   id: number;
   title: string;
   description: string;
-  deadline: string;
   budget: number;
   progress: number;
   CreatorId: number;
   activeStageId: number;
   activeStage: Stage;
-  Stages: Stage[];
+  Stages?: Stage[];
   Creator: User;
   Participants: User[];
   Documents: Asset[];
+  deadline: Date;
   createdAt: Date;
   updatedAt: Date;
 }
