@@ -1,20 +1,30 @@
-import { SharedModule } from '@/shared/shared.module';
-import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
+import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
+
 import { provideMockStore } from '@ngrx/store/testing';
 
+import { SharedModule } from '@/shared/shared.module';
+import { authStateMock } from '@/shared/testing/mocks';
+import { initialRolesState } from '../../store/role.reducer';
+
 import { RoleComponent } from './role.component';
+import { initialUsersState } from '@/modules/users/store/user.reducer';
 
 describe('RoleComponent', () => {
   let component: RoleComponent;
   let fixture: ComponentFixture<RoleComponent>;
+  const initialState = {
+    auth: authStateMock,
+    users: initialUsersState,
+    roles: initialRolesState
+  };
 
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
         declarations: [RoleComponent],
         imports: [RouterTestingModule, SharedModule],
-        providers: [provideMockStore({})]
+        providers: [provideMockStore({ initialState })]
       }).compileComponents();
     })
   );
