@@ -1,206 +1,58 @@
-import { Component, OnInit, OnDestroy, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
+import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Observable, Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
 
 import { Store, select } from '@ngrx/store';
-import { cloneDeep } from 'lodash';
+import { Observable } from 'rxjs';
+import { filter, tap } from 'rxjs/operators';
 
 import { AppState } from '@/core/reducers';
 import { isPrivileged } from '@/core/auth/store/auth.selectors';
-import { UsersDialogComponent } from '@/modules/users/components/dialog/users-dialog.component';
-import { User } from '@/modules/users';
-import { Role, Privilege } from '../../models';
-import { updateRoleRequested, changeIsEditingState } from '../../store/role.actions';
-import { PrivilegesDialogComponent } from '../privileges/dialog/privileges-dialog.component';
-import { selectIsEditing } from '../../store/role.selectors';
-import {
-  EDIT_PRIVILEGES,
-  COLUMN_NAMES,
-  COLUMN_LABELS,
-  ACTION_LABELS,
-  THEME_PALETTE,
-  MAT_BUTTON,
-  CONSTANTS,
-  RoutingDataConstants
-} from '@/shared/constants';
-import { DialogConfirmModel } from '@/shared/models/dialog/dialog-confirm.model';
-import { DialogDataModel } from '@/shared/models/dialog/dialog-data.model';
-import { DialogConfirmComponent } from '@/shared/components/dialogs/dialog-confirm/dialog-confirm.component';
-import { DialogWithTwoButtonModel } from '@/shared/models/dialog/dialog-with-two-button.model';
-import { DialogResultModel } from '@/shared/models/dialog/dialog-result.model';
-import { DialogType } from '@/shared/models';
-import { DynamicForm } from '@/shared/models';
-import { DialogService } from '@/shared/services';
+import { EDIT_PRIVILEGES, FORMCONSTANTS } from '@/shared/constants';
+import { updateRoleRequested, changeIsEditingState, roleRequested } from '../../store/role.actions';
+import { selectIsEditing, selectRoleDeepById } from '../../store/role.selectors';
+import { Role } from '../../models';
 
 @Component({
-  templateUrl: './role.component.html',
-  styleUrls: ['./role.component.scss'],
+  template: `
+    <templates-role-view
+      [item]="role$ | async"
+      [editForm]="editForm$ | async"
+      [formName]="formName"
+      [canEdit]="canEditRole$ | async"
+      (isEditing)="onFormStateChange($event)"
+      (saveChanges)="updateRole($event)"
+    ></templates-role-view>
+  `,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class RoleComponent implements OnInit, OnDestroy {
+export class RoleComponent implements OnInit {
   editForm$: Observable<boolean> = this.store$.pipe(select(selectIsEditing));
   canEditRole$: Observable<boolean> = this.store$.pipe(select(isPrivileged(EDIT_PRIVILEGES.ROLE)));
+  role$: Observable<Role>;
 
-  role: Role;
-  roleFormJson: DynamicForm;
-  roleInitial: Role;
-  roleFormValues: Role;
+  formName = FORMCONSTANTS.ROLE;
 
-  themePalette = THEME_PALETTE;
-  matButtonType = MAT_BUTTON;
-  columns = COLUMN_NAMES;
-  columnLabels = COLUMN_LABELS;
-  actionLabels = ACTION_LABELS;
-  displayedColumns: COLUMN_NAMES[] = [
-    COLUMN_NAMES.TITLE,
-    COLUMN_NAMES.VIEW,
-    COLUMN_NAMES.ADD,
-    COLUMN_NAMES.EDIT,
-    COLUMN_NAMES.DELETE
-  ];
-
-  private unsubscribe: Subject<void> = new Subject();
-
-  constructor(
-    private route: ActivatedRoute,
-    private store$: Store<AppState>,
-    private cdr: ChangeDetectorRef,
-    private dialogService: DialogService
-  ) {}
+  constructor(private route: ActivatedRoute, private store$: Store<AppState>) {}
 
   ngOnInit(): void {
-    this.roleFormJson = this.route.snapshot.data[RoutingDataConstants.FORM_JSON];
-    this.getRoleData();
-  }
+    const id: number = this.route.snapshot.params['id'];
 
-  roleFormValueChanges(formVal: Role): void {
-    this.roleFormValues = { ...this.roleFormValues, ...formVal };
-  }
-
-  getRoleData(): void {
-    this.role = cloneDeep(this.route.snapshot.data[RoutingDataConstants.ROLE]);
-    this.roleInitial = cloneDeep(this.route.snapshot.data[RoutingDataConstants.ROLE]);
-  }
-
-  addParticipantDialog(): void {
-    const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
-      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_USERS)
-    };
-
-    this.dialogService
-      .open(UsersDialogComponent, dialogDataModel, DialogType.MAX)
-      .afterClosed()
-      .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: DialogResultModel<User[]>) => {
-        if (result && result.success) {
-          const selectedUsers: User[] = result.model.filter(
-            (selectedUser) => !this.role.Users.some((user) => user.id === selectedUser.id)
-          );
-          if (selectedUsers?.length) {
-            this.role.Users = [...this.role.Users, ...selectedUsers];
-            this.cdr.detectChanges();
-          }
+    this.role$ = this.store$.pipe(
+      select(selectRoleDeepById(id)),
+      tap((role) => {
+        if (!role) {
+          this.store$.dispatch(roleRequested({ id }));
         }
-      });
-
-    // const userC = dialogRef.componentInstance.usersComponent;
-
-    // dialogRef
-    //   .afterOpened()
-    //   .pipe(takeUntil(this.unsubscribe), skipUntil(userC.loading$), delay(300))
-    //   .subscribe(() => {
-    //     userC.users
-    //       .filter((user) => this.role.Users.some((rUser) => rUser.id === user.id))
-    //       ?.forEach((selectedParticipant) => {
-    //         userC.selection.select(selectedParticipant);
-    //       });
-    //   });
+      }),
+      filter((role) => !!role)
+    );
   }
 
-  addPrivilegeDialog(): void {
-    const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
-      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_PRIVILEGES)
-    };
-
-    this.dialogService
-      .open(PrivilegesDialogComponent, dialogDataModel, DialogType.STANDART)
-      .afterClosed()
-      .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: DialogResultModel<Privilege[]>) => {
-        if (result && result.success) {
-          const selectedPrivileges: Privilege[] = result.model
-            .filter(
-              (selectedPrivilege) => !this.role.Privileges.some((rPrivilege) => rPrivilege.id === selectedPrivilege.id)
-            )
-            ?.map((selectedPrivilege) => {
-              return {
-                ...selectedPrivilege,
-                RolePrivilege: {
-                  add: false,
-                  view: false,
-                  edit: false,
-                  delete: false
-                }
-              };
-            });
-
-          if (selectedPrivileges?.length) {
-            this.role.Privileges = [...this.role.Privileges, ...selectedPrivileges];
-            this.cdr.detectChanges();
-          }
-        }
-      });
-
-    // const privilegesC = dialogRef.componentInstance.privilegesComponent;
-
-    // dialogRef
-    //   .afterOpened()
-    //   .pipe(takeUntil(this.unsubscribe), skipUntil(privilegesC.isLoading$), delay(300))
-    //   .subscribe(() => {
-    //     privilegesC.privileges
-    //       .filter((privilege) => this.role.Privileges.some((rPrivilege) => rPrivilege.id === privilege.id))
-    //       ?.forEach((selectedPrivilege) => {
-    //         privilegesC.selection.select(selectedPrivilege);
-    //       });
-    //   });
+  onFormStateChange(isEditing: boolean): void {
+    this.store$.dispatch(changeIsEditingState({ isEditing }));
   }
 
-  removePrivilege(privilegeId: number): void {
-    this.role = { ...this.role, Privileges: this.role.Privileges.filter((privilege) => privilege.id !== privilegeId) };
-  }
-
-  removeUser(userId: number): void {
-    this.role = { ...this.role, Users: this.role.Users.filter((user) => user.id !== userId) };
-  }
-
-  onClickEdit(): void {
-    this.store$.dispatch(changeIsEditingState({ isEditing: true }));
-    this.displayedColumns = [...this.displayedColumns, COLUMN_NAMES.ACTIONS];
-  }
-
-  onClickCancelEdit(): void {
-    this.store$.dispatch(changeIsEditingState({ isEditing: false }));
-    this.role = cloneDeep(this.roleInitial);
-    this.disableEdit();
-  }
-
-  disableEdit(): void {
-    this.displayedColumns = this.displayedColumns.filter((col) => col !== COLUMN_NAMES.ACTIONS);
-  }
-
-  updateRole(): void {
-    const dialogModel: DialogConfirmModel = new DialogConfirmModel(CONSTANTS.TEXTS_UPDATE_ROLE_CONFIRM);
-    const dialogDataModel: DialogDataModel<DialogConfirmModel> = { dialogModel };
-
-    this.dialogService.confirm(DialogConfirmComponent, dialogDataModel, () => {
-      this.store$.dispatch(updateRoleRequested({ role: { ...this.role, ...this.roleFormValues } }));
-      this.disableEdit();
-    });
-  }
-
-  ngOnDestroy(): void {
-    this.unsubscribe.next();
-    this.unsubscribe.complete();
+  updateRole(role: Role): void {
+    this.store$.dispatch(updateRoleRequested({ role }));
   }
 }
